# shared/api_spec.yaml
openapi: 3.0.3
info:
  title: DPS Dashboard API
  description: >
    Deshapriya Park Durga Puja Club Management Dashboard API.
    All endpoints require authentication unless marked as public.
    Role enforcement is server-side on every route.
  version: 1.0.0
  contact:
    name: DPS Dashboard Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: http://{vps_ip}/api
    description: LunaNode VPS
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: HTTP-only JWT session cookie set by NextAuth
  schemas:
...
    MembershipStatus:
      type: string
      enum: [PENDING_APPROVAL, PENDING_PAYMENT, ACTIVE, EXPIRED, SUSPENDED]
...
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    ApprovalEntityType:
      type: string
      enum: [TRANSACTION, MEMBER_ADD, MEMBER_EDIT, MEMBER_DELETE, MEMBERSHIP]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 8
        newPassword:
          type: string
          minLength: 8

    # Members
...
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^\+91\d{10}$'
        address:
          type: string
    MemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          description: "Format: DPC-YYYY-NNNN-00"
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        membershipStatus:
          $ref: '#/components/schemas/MembershipStatus'
        joinedAt:
          type: string
          format: date-time
        subMembers:
          type: array
          items:
            $ref: '#/components/schemas/SubMemberResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SubMemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          description: "Format: DPC-YYYY-NNNN-SS (01-03)"
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        relation:
          type: string
        canLogin:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # Memberships
...
    ApprovalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entityType:
          $ref: '#/components/schemas/ApprovalEntityType'
        entityId:
          type: string
          format: uuid
        action:
          type: string
        previousData:
          type: object
          nullable: true
        newData:
          type: object
          nullable: true
        requestedBy:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
        status:
          $ref: '#/components/schemas/ApprovalStatus'
        reviewedBy:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    ActivityLogResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
        action:
          type: string
        description:
          type: string
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time

    # Dashboard
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
security:
  - cookieAuth: []
paths:
...
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change password (forced for temp passwords)
      description: "Users with isTempPassword=true must call this before accessing any dashboard route."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== MEMBERS ====================
...
  /members/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Members]
      summary: Get single member
      description: "Admin/Operator: any member. Member: own record only."
      responses:
        '200':
          description: Member details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Members]
      summary: Update member
      description: >
        Admin: updates directly.
        Operator: creates Approval record (MEMBER_EDIT) with previousData/newData.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
      responses:
        '200':
          description: Member updated (or approval pending for operator)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MemberResponse'
                  - $ref: '#/components/schemas/ApprovalResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Members]
      summary: Delete member
      description: >
        Admin: deletes directly.
        Operator: creates Approval record (MEMBER_DELETE).
      responses:
        '200':
          description: Member deleted (or approval pending for operator)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /approvals:
    get:
      tags: [Approvals]
      summary: List pending approvals
      description: "Admin only. Returns all PENDING approvals."
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: entityType
          in: query
          schema:
            $ref: '#/components/schemas/ApprovalEntityType'
      responses:
        '200':
          description: Paginated approvals
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ApprovalResponse'
        '403':
          description: Not admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /activity-log:
    get:
      tags: [Logs]
      summary: System-wide activity log
      description: "Admin: full access. Operator: read-only."
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: action
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paginated activity log entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ActivityLogResponse'

  # ==================== RECEIPTS ====================
...